==> 2_ncd_tables_gee.Rmd <==
---
title: "DO ART NCD Tables & Figures"
author: ""
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  word_document:
    reference_docx: word_styles.docx

---

```{r setup, include = F, echo = F, warning = F, message = F}

rm(list=ls())

# Load packages 
package.list <- c("plyr", "dplyr", "tidyr", "magrittr", "data.table", "ggplot2",
                  "janitor", "tables", "flextable", "ggpubr", # table formatting
                  "geepack", "sandwich") # regression analysis   
invisible(lapply(package.list, library, character.only = TRUE))

# Load data and functions
dir <- "C:/Users/msahu/OneDrive - UW/Documents/Research/DGH+MGH/DO_ART_NCD/"
in_dir <- paste0(dir, "0_data/3_cleaned_data/")
source(paste0(dir, "2_code/gee_functions.R"))

# Subset to South Africa only!!
# NOTE: we combine the community and hybrid arm (community follow-up), compared to clinic
ncd_merged_subset <- readRDS(paste0(in_dir, "ncd_merged_subset.rds")) %>% 
  filter(country == "South Africa") %>% 
  mutate(site = factor (site, levels = c("HSRC", "AHRI")))

# Number plugging: % individuals in a household with multiple participants?
total_n <- ncd_merged_subset %>% nrow()
unique_hh <- length(unique(ncd_merged_subset$hhid))

```

## Main paper tables (South Africa only)

+ Table 1: Descriptive statistics of cardiovascular risk in the South Africa site of the DO ART study, by trial arm and gender
+ Table 2: Poisson + linear regressions for community follow-up compared with clinic follow-up
+ Table 3: Poisson + linear regressions with VS as predictor, and including CD4+
+ Table 4: Poisson + linear regressions with VS as predictor, stratified by treatment arm; with interaction term for VS

&nbsp;

Notes:

+ Baseline models include study site only
+ Adjusted models include: study site, gender, age group, smoking (binary; otherwise we get small group problems), education
&nbsp;


### Page Break

## Table 1: Descriptive statistics of cardiovascular risk in the South Africa site of the DO ART study, by trial arm and gender (N = 1010)

```{r, echo = F, warning = F, message = F, ft.split = T, ft.keepnext = F, results = 'asis'}
# Vars to include
table_vars <- data.frame(
  
  var_name = c(
    A_vars, 
    B_vars, 
    C_vars
  ),
  
  var_label = c(A_var_labels, B_var_labels, C_var_labels)) %>% 
  
  mutate( 
    
    section = case_when(
      var_name %in% c(A_vars) ~ "A",
      var_name %in% c(B_vars) ~ "B",
      var_name %in% c(C_vars) ~ "C")
  )


# Descriptive Proportion Table 
col_headings_freq <- c(
  "Category",
  "Clinic follow-up \n n (%)", 
  "Community follow-up \n n (%)",
  "Women \n n (%)",
  "Men \n n (%)",
  "Total \n n (%)"
)

for (i in unique(headings$section)) {
  
  table_subset <- table_vars[table_vars$section==i, ]
  
  tableDF <- data.frame(
        variable = character(),
        category = character(),
        clinic_arm = character(),
        community_arm = character(),
        Women = character(),
        Men = character(),
        Total = character()
    )
  
  for (v in unique(table_subset$var_name)) {
      
    data = ncd_merged_subset
    
    temp1 <- data %>% 
      
      # Format table using Janitor package 
      tabyl(!!sym(v), armR) %>% 
      adorn_totals(where = "col") %>% 
      adorn_percentages(denominator = "col", na.rm = T) %>% 
      adorn_pct_formatting(digits = 0, affix_sign = TRUE) %>% 
      adorn_ns(position = "front") %>%
      as.data.frame() %>% 
      
      # Clean up headers
      dplyr::rename(category = 1,
             clinic_arm = `Clinic follow-up`,
             community_arm = `Community follow-up`) %>% 
      mutate(variable = table_vars[table_vars$var_name == v, "var_label"] ) %>% 
      select(variable, category, clinic_arm, community_arm, Total)
    
    if (v != "genderR") {
    
    temp2 <- data %>% 
      
      # Format table using Janitor package 
      tabyl(!!sym(v), genderR) %>% 
      adorn_totals(where = "col") %>% 
      adorn_percentages(denominator = "col") %>% 
      adorn_pct_formatting(digits = 0, affix_sign = TRUE) %>% 
      adorn_ns(position = "front") %>%
      as.data.frame() %>% 
      
      # Clean up headers
      select(Women, Men)
    
    }
    
    if (v == "genderR") {
    
    temp2 <- data.frame(
      Women = "--",
      Men = "--"
    )
    
    }
  
  temp <- cbind(temp1, temp2) %>% 
    select(variable, category, clinic_arm, community_arm, Women, Men, Total)
      
  tableDF <- rbind(tableDF, temp)
  
  }
  
  # Flextable

ft <- flextable(tableDF) %>%
   
  delete_part(part = "header") %>%
  add_header_row(values = c(headings[headings$section == i, "title"], col_headings_freq))  %>%     
  align(i = 1, part = "header", align = "center") %>%
  merge_v(j = "variable") %>%
  width(width = 1.4) %>%
  theme_vanilla() %>% 
  vline(j = c(2, 4, 6), border = fp_border_default(), part = "all")

flextable_to_rmd(ft)  

rm(temp, temp1, temp2)  
}

```


### Page Break

## Table 2. Cardiovascular risk for community follow-up care versus clinic follow-up care, South Africa (N = 1010).

```{r, echo = F, warning = F, message = F, ft.split = T, ft.keepnext = F, results = 'asis'}

df = ncd_merged_subset
var_of_interest = "armR"
var_to_pull = "armRCommunity follow-up"
ext = F

# POISSON 
ft <- get_regression_results(df = df, var_of_interest = var_of_interest, var_to_pull = var_to_pull, family = "poisson", ext = ext)
ft

# LINEAR 
ft <- get_regression_results(df = df, var_of_interest = var_of_interest, var_to_pull = var_to_pull, family = "gaussian", ext = ext)
ft
```

### Page Break

## Table 3: Cardiovascular risk among people virally suppressed versus those not virally suppressed at endline and controlling for CD4+ count, in South Africa(N = 983).

```{r, echo = F, warning = F, message = F, ft.split = T, ft.keepnext = F, results = 'asis'}

df <- ncd_merged_subset %>% filter(!is.na(exit_viral_load_result))
var_of_interest = "exit_viral_load_suppressed"
var_to_pull = "exit_viral_load_suppressedTRUE"
ext = T 

# POISSON 
ft <- get_regression_results(df = df, var_of_interest = var_of_interest, var_to_pull = var_to_pull, family = "poisson", ext = ext)
ft

# LINEAR 
ft <- get_regression_results(df = df, var_of_interest = var_of_interest, var_to_pull = var_to_pull, family = "gaussian", ext = ext)
ft

```


### PAGE BREAK

## Table 4: Cardiovascular risk for persons virally suppressed versus not suppressed at exit for clinic and community follow-up care, South Africa (N = 983). 
```{r, echo = F, warning = F, message = F, ft.split = T, ft.keepnext = F, results = 'asis'}
df <- ncd_merged_subset %>% filter(!is.na(exit_viral_load_suppressed))
var_of_interest = "exit_viral_load_suppressed"
var_to_pull = "exit_viral_load_suppressedTRUE"
family = "poisson"
strat_var = "armR"
strat_var_to_pull = "armRCommunity follow-up"

# POISSON 
ft <- get_reg_strat_results(df = df, family = "poisson",
                             var_of_interest = var_of_interest, var_to_pull = var_to_pull,  
                             strat_var = strat_var, strat_var_to_pull = strat_var_to_pull)
ft

# LINEAR
ft <- get_reg_strat_results(df = df, family = "gaussian",
                             var_of_interest = var_of_interest, var_to_pull = var_to_pull,  
                             strat_var = strat_var, strat_var_to_pull = strat_var_to_pull)
ft
```
